# openapi.yaml
openapi: 3.0.3
info:
  title: NFL Analytics Platform API
  version: 7.0.0
  description: |
    Enterprise-grade NFL data extraction, machine learning analytics, and real-time processing platform.
    
    ## Features
    - **19 NFL Datasets**: Complete access to play-by-play, player stats, schedules, and more
    - **Machine Learning**: Fantasy predictions, player consistency analysis, breakout candidates
    - **Real-time Processing**: Live game updates, WebSocket streaming, event processing
    - **Production Ready**: Docker deployment, monitoring, CI/CD integration
    
    ## Quick Start
    1. Extract NFL data: `POST /api/v1/datasets/{dataset}/extract`
    2. Train ML models: `POST /api/v1/analytics/train`
    3. Get predictions: `POST /api/v1/analytics/predict`
    4. Monitor live games: `GET /api/v1/realtime/games/live`
    
  contact:
    name: NFL Analytics Platform
    url: https://github.com/anotherjson/another-nfl
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://nfl-analytics.example.com/api/v1
    description: Production server
  - url: ws://localhost:8765
    description: WebSocket server for real-time updates
...
paths:
...
  /datasets/{dataset}:
    get:
      tags: [datasets]
      summary: Get dataset configuration
      description: Retrieve configuration details for specific NFL dataset
      parameters:
        - $ref: '#/components/parameters/DatasetName'
      responses:
        200:
          description: Dataset configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetConfig'
        404:
          $ref: '#/components/responses/NotFound'
  /datasets/{dataset}/extract:
    post:
      tags: [datasets]
      summary: Extract NFL dataset
      description: |
        Extract NFL data for specified dataset and year. Supports both year-based 
        datasets (pbp, weekly, seasonal) and static datasets (team_desc, players).
      parameters:
        - $ref: '#/components/parameters/DatasetName'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtractionRequest'
            examples:
              yearly_dataset:
                summary: Extract play-by-play data for 2023
                value:
                  year: 2023
                  validate: true
                  save_to_disk: true
              static_dataset:
                summary: Extract team descriptions
                value:
                  validate: true
                  save_to_disk: true
      responses:
        200:
          description: Extraction completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractionResult'
        400:
          $ref: '#/components/responses/BadRequest'
        422:
          $ref: '#/components/responses/ValidationError'
...
  /analytics/train:
    post:
      tags: [analytics]
      summary: Train ML model
      description: |
        Train machine learning models on NFL data. Supports fantasy prediction,
        consistency analysis, and team strength models.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingRequest'
            examples:
              fantasy_model:
                summary: Train fantasy prediction model for RBs
                value:
                  model_type: fantasy
                  position: RB
                  years: [2020, 2021, 2022, 2023]
                  algorithm: random_forest
      responses:
        200:
          description: Model training completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingResult'
        202:
          description: Training started (async)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncTaskStatus'
  /analytics/predict:
    post:
      tags: [analytics]
      summary: Generate predictions
      description: Generate predictions using trained ML models
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionRequest'
      responses:
        200:
          description: Predictions generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionResult'
...
components:
  parameters:
    DatasetName:
      name: dataset
      in: path
      required: true
      description: NFL dataset name
      schema:
        type: string
        enum: [pbp, weekly, seasonal, weekly_rosters, seasonal_rosters, schedules, team_desc, officials, combine, draft_picks, qbr, weekly_pfr, seasonal_pfr, injuries, depth_charts, snap_counts, ftn_data, ngs_data, players]
      example: pbp
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
  schemas:
...
    Dataset:
      type: object
      required: [name, description, start_year, requires_year, data_type]
      properties:
        name:
          type: string
          description: Dataset identifier
          example: pbp
        description:
          type: string
          description: Human-readable description
          example: "Play-by-play data with detailed game information"
        start_year:
          type: integer
          nullable: true
          description: Earliest available year (null for non-year datasets)
          example: 1999
        requires_year:
          type: boolean
          description: Whether this dataset requires a year parameter
          example: true
        data_type:
          type: string
          description: Category of data
          enum: [game_level, player_weekly, player_seasonal, team_level, static]
          example: game_level
        estimated_size_mb:
          type: number
          description: Estimated file size in MB for recent years
          example: 150.5
    DatasetConfig:
      allOf:
        - $ref: '#/components/schemas/Dataset'
        - type: object
          properties:
            validation:
              type: object
              properties:
                required_columns:
                  type: array
                  items:
                    type: string
                expected_size_mb:
                  type: number
            extraction:
              type: object
              properties:
                timeout_seconds:
                  type: integer
                retry_attempts:
                  type: integer
            output:
              type: object
              properties:
                file_format:
                  type: string
                  enum: [parquet, csv, json]
                compression:
                  type: string
                  enum: [snappy, gzip, none]

    # Extraction Models
    ExtractionRequest:
      type: object
      properties:
        year:
          type: integer
          minimum: 1936
          maximum: 2030
          description: Year to extract (for year-based datasets)
          example: 2023
        validate:
          type: boolean
          default: true
          description: Enable data validation
        save_to_disk:
          type: boolean
          default: true
          description: Save extracted data to parquet file
        verbose:
          type: boolean
          default: false
          description: Enable verbose logging
    ExtractionResult:
      type: object
      required: [dataset, success, rows, columns, extraction_time_seconds]
      properties:
        dataset:
          type: string
          example: pbp
        year:
          type: integer
          nullable: true
          example: 2023
        success:
          type: boolean
          example: true
        rows:
          type: integer
          description: Number of rows extracted
          example: 45287
        columns:
          type: integer
          description: Number of columns
          example: 372
        file_size_mb:
          type: number
          description: File size in MB (if saved)
          example: 156.8
        extraction_time_seconds:
          type: number
          description: Time taken for extraction
          example: 23.4
        validation_passed:
          type: boolean
          description: Whether validation checks passed
          example: true
        file_path:
          type: string
          nullable: true
          description: Path to saved file (if applicable)
          example: "data/pbp/2023/etl_date=2025-07-22/data.parquet"
...
    Position:
      type: string
      enum: [QB, RB, WR, TE, K, DEF, all]
      example: RB
...
    TrainingRequest:
      type: object
      required: [model_type]
      properties:
        model_type:
          type: string
          enum: [fantasy, consistency, strength, breakout]
          example: fantasy
        position:
          $ref: '#/components/schemas/Position'
        years:
          type: array
          items:
            type: integer
          example: [2020, 2021, 2022, 2023]
        algorithm:
          type: string
          enum: [random_forest, gradient_boosting, linear_regression, auto]
          default: auto
        save_model:
          type: boolean
          default: true
        model_name:
          type: string
          description: Custom name for saved model
    TrainingResult:
      type: object
      properties:
        model_type:
          type: string
        position:
          $ref: '#/components/schemas/Position'
        algorithm:
          type: string
        training_samples:
          type: integer
        features_used:
          type: integer
        cv_score:
          type: number
          description: Cross-validation score
        mae:
          type: number
          description: Mean Absolute Error
        rmse:
          type: number
          description: Root Mean Square Error
        feature_importance:
          type: object
          additionalProperties:
            type: number
        model_path:
          type: string
          nullable: true
          description: Path to saved model file
    PredictionRequest:
      type: object
      required: [model_type]
      properties:
        model_type:
          type: string
          enum: [fantasy, consistency, strength]
        model_path:
          type: string
          nullable: true
          description: Path to specific model file
        position:
          $ref: '#/components/schemas/Position'
        player_name:
          type: string
          description: Specific player to predict for
        week:
          type: integer
          minimum: 1
          maximum: 18
          description: Week to predict for
        data:
          type: object
          description: Custom player data for prediction
    PredictionResult:
      type: object
      properties:
        model_type:
          type: string
        position:
          $ref: '#/components/schemas/Position'
        predictions:
          type: array
          items:
            type: object
            properties:
              player_name:
                type: string
              predicted_points:
                type: number
              confidence_lower:
                type: number
              confidence_upper:
                type: number
              prediction_rank:
                type: integer
...
    AsyncTaskStatus:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        progress_percent:
          type: number
          minimum: 0
          maximum: 100
        estimated_completion:
          type: string
          format: date-time
          nullable: true

    # Error Models
    Error:
      type: object
      required: [status, error]
      properties:
        status:
          type: string
          enum: [error]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: INVALID_DATASET
            message:
              type: string
              example: "Dataset 'invalid_name' not found"
            details:
              type: object
              additionalProperties: true
        metadata:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            version:
              type: string
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            validation_errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  value:
                    nullable: true
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - bearerAuth: []
  - apiKeyAuth: []

# WebSocket specification (using AsyncAPI concepts in OpenAPI)
...